+++ src/Entities/EntityDecorator.php
@@ -4,6 +4,7 @@ namespace Bozboz\Jam\Entities;
 
 use Bozboz\Admin\Base\ModelAdminDecorator;
 use Bozboz\Admin\Fields\BelongsToManyField;
+use Bozboz\Admin\Fields\DateTimeField;
 use Bozboz\Admin\Fields\HiddenField;
 use Bozboz\Admin\Fields\TextField;
 use Bozboz\Admin\Fields\URLField;
@@ -46,6 +47,44 @@ class EntityDecorator extends ModelAdminDecorator
 		$columns->prepend($this->getPreviewLink($instance), 'Name');
 		$columns->put('', $this->getLockState($instance));
 
+		// $linkText = '<i class="fa fa-external-link"></i>';
+		// $path = $instance->canonical_path;
+
+		// switch ($instance->status) {
+		// 	case Revision::PUBLISHED:
+		// 		$publishedAt = $instance->currentRevision->formatted_published_at;
+		// 		$user = $instance->currentRevision->username;
+		// 		$statusLabel = "<small><abbr title='{$publishedAt} by {$user}'>Published</abbr></small>";
+		// 	break;
+
+		// 	case Revision::SCHEDULED:
+		// 		$publishedAt = $instance->currentRevision->formatted_published_at;
+		// 		$user = $instance->currentRevision->username;
+		// 		$statusLabel = "<small><abbr title='{$publishedAt} by {$user}'>Scheduled</abbr></small>";
+		// 	break;
+
+		// 	case Revision::PUBLISHED_WITH_DRAFTS:
+		// 		$publishedAt = $instance->latestRevision()->created_at->format('d-m-Y H:i');
+		// 		$user = $instance->latestRevision()->username;
+		// 		$statusLabel = "<small><abbr title='{$publishedAt} by {$user}'>Has Draft</abbr></small>";
+
+		// 		$linkText = 'preview <i class="fa fa-external-link"></i>';
+		// 		$path = $instance->canonical_path . '?p=' . md5(date('ymd'));
+		// 	break;
+
+		// 	default:
+		// 		$statusLabel = null;
+		// 		$linkText = 'preview <i class="fa fa-external-link"></i>';
+		// 		$path = $instance->canonical_path . '?p=' . md5(date('ymd'));
+		// 	break;
+		// }
+
+		// $columns = collect($this->getCustomColumns($instance));
+		// $columns->prepend(
+		// 	$this->getLabel($instance) . ( $path
+		// 		? '&nbsp;&nbsp;<a href="'.url($path).'" target="_blank" title="Go to '.$this->getLabel($instance).'">'.$linkText.'</a>'
+		// 		: null
+		// ), 'Name')->put('Status', $statusLabel);
 		return $columns->all();
 	}
 
@@ -99,7 +138,12 @@ class EntityDecorator extends ModelAdminDecorator
 				'label' => 'Restrict visibility by role',
 				'help_text' => 'Leave blank for full public access'
 			]),
-			$canEditStatus ? new PublishField('status', $instance) : null,
+			$canEditStatus
+				? new DateTimeField('currentRevision[published_at]', [
+					'label' => 'Published At',
+					'help_text' => 'If you enter a date in the future, this will be hidden from the site until that date is reached.',
+				])
+				: new HiddenField('currentRevision[published_at]'),
 		]));
 
 		return $fields->merge($this->getTemplateFields($instance))->all();
@@ -192,7 +236,7 @@ class EntityDecorator extends ModelAdminDecorator
 			$query->whereTypeAlias($this->type->alias);
 		});
 
-		$query->ordered();
+		$query->ordered()->with('revisions');
 	}
 
 	public function findInstance($id)