+++ src/Fields/BelongsToMany.php
@@ -64,11 +64,24 @@ abstract class BelongsToMany extends BelongsTo
         $entity->setAttribute($this->getInputName(), $this->relation($value)->getRelatedIds()->all());
     }
 
+    public function injectDiffValue(Entity $entity, Revision $revision)
+    {
+        $value = $revision->fieldValues->where('key', $this->name)->first() ?: new Value(['key' => $this->name]);
+        $this->injectValue($entity, $value);
+        $entity->setAttribute(
+            $value->key,
+            $entity->getAttribute($value->key)->map(function($entity) {
+                return $entity->name;
+            })->implode("\n")
+        );
+        return $value;
+    }
+
     public function relation(Value $value)
     {
         $pivot = $this->getPivot();
         $query = $value->belongsToMany($this->getRelationModel(), $pivot->table, $pivot->foreign_key, $pivot->other_key);
-        return $query;
+        return $query->active();
     }
 
     public function saveValue(Revision $revision, $value)