+++ src/Fields/Password.php
@@ -5,6 +5,7 @@ namespace Bozboz\Jam\Fields;
 use Bozboz\Admin\Fields\PasswordField;
 use Bozboz\Jam\Entities\Entity;
 use Bozboz\Jam\Entities\EntityDecorator;
+use Bozboz\Jam\Entities\Revision;
 use Bozboz\Jam\Entities\Value;
 
 class Password extends Field
@@ -18,4 +19,11 @@ class Password extends Field
             'help_text' => $this->help_text,
 		]);
 	}
+
+    public function injectDiffValue(Entity $entity, Revision $revision)
+    {
+        $value = $revision->fieldValues->where('key', $this->name)->first() ?: new Value(['key' => $this->name]);
+        $entity->setAttribute($value->key, str_repeat('*', strlen($value->value)));
+        return $value;
+    }
 }