+++ src/Fields/SortedBelongsToMany.php
@@ -28,7 +28,8 @@ abstract class SortedBelongsToMany extends BelongsToMany
 
     public function relation(Value $value)
     {
-        return parent::relation($value)->withPivot('sorting')->orderBy('sorting');
+        return parent::relation($value)
+            ->withPivot('sorting')->orderBy($this->getPivot()->table . '.sorting');
     }
 
     public function saveValue(Revision $revision, $value)
@@ -51,7 +52,7 @@ abstract class SortedBelongsToMany extends BelongsToMany
 
     public function duplicateValue(Value $oldValue, Value $newValue)
     {
-        $syncData = $this->relation($oldValue)->withPivot('sorting')->pluck($this->getPivot()->foreign_key, 'sorting')->toArray();
+        $syncData = $this->relation($oldValue)->withPivot('sorting')->pluck($this->getPivot()->other_key, 'sorting')->toArray();
         $this->relation($newValue)->sync($syncData);
     }
 }