+++ src/Http/Controllers/EntityController.php
@@ -9,7 +9,7 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class EntityController extends Controller
 {
-    private $repository;
+    protected $repository;
 
     public function __construct(EntityRepository $repository)
     {