+++ changelog.md
@@ -1,6 +1,39 @@
 # Jam Changelog
 
-## Version 1.0.0
+## Version 1.1.0 (2017-01-09)
+
+-   Rearrange publishing/scheduling feature for better usability
+-   Allow draft revisions to be created on published entities
+-   Allow previewing unpublished entities/draft revisions
+-   Add revision diff feature
+
+## Version 1.0.6 (Future)
+
+-   Add validation to fields to prevent duplicate names on templates
+
+## Version 1.0.5 (2017-01-09)
+
+-   Fix value duplication for belongs to fields
+
+## Version 1.0.4 (2016-12-23)
+
+-   Exclude inactive entities from BelongsToMany field query
+
+## Version 1.0.3 (2016-12-22)
+
+-   Make repo in front end controller protected rather than private
+-   Fix user field default value and make defaulting to logged in user optional
+-   Automatically inject values on new collection of entities if values have been eager loaded 
+
+## Version 1.0.2 (2016-12-21)
+
+-   Fix nested entity list redirect 
+
+## Version 1.0.1 (2016-12-16)
+
+-   Fix entity list save redirect for nested lists
+
+## Version 1.0.0 (2016-12-16)
 
 -   Add max_uses to templates
 -   Don't make slugs unique